Include repository.directory in generated package.json
The repository.directory field in package.json allows machines (and
humans) to know where to look in the mono-repository to find the package.

Including this field is of particular use to tools like Dependabot, when
surfacing resources such as changelogs.

As we use init-package-json to generate the package.json, and it does
not support the directory field, we must edit the generated package.json
ourselves. Fortunately, this is fairly straightforward.